+++ CHANGELOG.md
@@ -7,3 +7,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Create data fixture to init application database
+- Create console commands: `app:user:new`, `app:user:list`, `app:user:view`, `app:user:assign-team`